refactor: rename project to "Code Shelf" and update version to 0.0.1
- Updated README.md to reflect the new project name and features.
- Changed version in package.json and package-lock.json to 0.0.1.
- Removed deprecated 404.html file.
- Updated FooterMenu and Header components to use BASE_URL for path matching.
- Modified navItems in menu.tsx to include new paths and icons.
- Deleted unused cssTips.ts file and created a new tips.ts file for general tips.
- Removed TSTips page and replaced it with a new Tips page for general web development tips.
- Created UsefulLinks page to provide a collection of useful resources.
- Updated router paths to reflect new structure and removed old paths.
