--- src/data/cssTips.ts
@@ -1,70 +1,0 @@
-import { cssExamples } from './cssExamples';
-
-export const tips = [
-  {
-    title: 'Use Flexbox for Centering',
-    description: 'The easiest way to center content both horizontally and vertically',
-    code: cssExamples.flexboxCenter,
-    category: 'Layout',
-  },
-  {
-    title: 'CSS Grid for Responsive Layouts',
-    description: 'Create flexible, responsive layouts with minimal code',
-    code: cssExamples.gridLayout,
-    category: 'Layout',
-  },
-  {
-    title: 'Custom Scrollbars',
-    description: 'Style scrollbars for better user experience',
-    code: cssExamples.customScrollbar,
-    category: 'Styling',
-  },
-  {
-    title: 'Gradient Text',
-    description: 'Apply gradients to text for modern design effects',
-    code: cssExamples.textGradient,
-    category: 'Styling',
-  },
-  {
-    title: 'Smooth Hover Effects',
-    description: 'Add smooth transitions for interactive elements',
-    code: cssExamples.hoverEffects,
-    category: 'Interactions',
-  },
-  {
-    title: 'Responsive Images',
-    description: 'Ensure images scale properly across all devices',
-    code: cssExamples.responsiveImages,
-    category: 'Responsive',
-  },
-  {
-    title: 'Dark Mode Support',
-    description: 'Implement automatic dark mode detection',
-    code: cssExamples.darkMode,
-    category: 'Accessibility',
-  },
-  {
-    title: 'CSS Animations',
-    description: 'Create smooth animations with staggered children',
-    code: cssExamples.animations,
-    category: 'Animations',
-  },
-  {
-    title: 'Accessibility Best Practices',
-    description: 'Ensure your CSS supports screen readers and keyboard navigation',
-    code: cssExamples.accessibility,
-    category: 'Accessibility',
-  },
-  {
-    title: 'Performance Optimizations',
-    description: 'Optimize CSS for better performance',
-    code: cssExamples.performance,
-    category: 'Performance',
-  },
-  {
-    title: 'Modern CSS Features',
-    description: 'Leverage the latest CSS features for better code',
-    code: cssExamples.modernCSS,
-    category: 'Modern',
-  },
-];
