+++ src/data/tips.ts
@@ -1,0 +1,20 @@
+export const tips = [
+  {
+    title: 'Test title',
+    description: 'Test description',
+    code: 'test code',
+    category: 'Css',
+  },
+  {
+    title: 'Test title',
+    description: 'Test description',
+    code: 'test code',
+    category: 'HTML',
+  },
+  {
+    title: 'Test title',
+    description: 'Test description',
+    code: 'test code',
+    category: 'Java Script',
+  },
+];
