--- src/pages/CSSTips/constants.tsx
@@ -1,12 +1,0 @@
-import { Code, Info, Brush, Palette, Lightbulb, TrendingUp, AutoAwesome, CheckCircle } from '@mui/icons-material';
-
-export const categories = [
-  { name: 'All Tips', icon: <Code /> },
-  { name: 'Layout', icon: <Brush /> },
-  { name: 'Styling', icon: <Palette /> },
-  { name: 'Interactions', icon: <AutoAwesome /> },
-  { name: 'Responsive', icon: <TrendingUp /> },
-  { name: 'Accessibility', icon: <CheckCircle /> },
-  { name: 'Performance', icon: <Lightbulb /> },
-  { name: 'Modern', icon: <Info /> },
-];
