+++ src/pages/Home/index.tsx
@@ -2,6 +2,9 @@ import React from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 import { Box, Typography, Card, CardContent, CardActions, Button, Paper, useTheme } from '@mui/material';
 import { features } from '../../data/features';
+import { PATH_EXAMPLES, PATH_SHOWCASE } from '../../router/path';
+
+const BASE_URL = import.meta.env.BASE_URL;
 
 export const Home: React.FC = () => {
   const theme = useTheme();
@@ -35,7 +38,7 @@ export const Home: React.FC = () => {
         <Box sx={{ display: 'flex', gap: 2, justifyContent: 'center', flexWrap: 'wrap' }}>
           <Button
             component={RouterLink}
-            to='/showcase'
+            to={`${BASE_URL}${PATH_SHOWCASE}`}
             variant='contained'
             size='large'
             sx={{
@@ -50,7 +53,7 @@ export const Home: React.FC = () => {
           </Button>
           <Button
             component={RouterLink}
-            to='/examples'
+            to={`${BASE_URL}${PATH_EXAMPLES}`}
             variant='outlined'
             size='large'
             sx={{
