+++ src/pages/Tips/constants.tsx
@@ -1,0 +1,8 @@
+import { Code, Css, Javascript, Html } from '@mui/icons-material';
+
+export const categories = [
+  { name: 'All Tips', icon: <Code /> },
+  { name: 'HTML', icon: <Html /> },
+  { name: 'Css', icon: <Css /> },
+  { name: 'Java Script', icon: <Javascript /> },
+];
