+++ src/router/path.ts
@@ -1,6 +1,6 @@
 export const PATH_ROOT = '';
-export const PATH_TS_TIPS = 'ts-tips';
+export const PATH_TIPS = 'tips';
 export const PATH_SHOWCASE = 'showcase';
 export const PATH_EXAMPLES = 'examples';
-export const PATH_CSS_TIPS = 'css-tips';
+export const PATH_USEFUL_LINKS = 'useful-links';
 export const PATH_DESIGN_SYSTEM = 'design-system';
