+++ src/router/router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from 'react-router-dom';
-import { CodeShowcase, CSSTips, DesignSystem, Examples, Home, TSTips } from '../pages';
-import { PATH_CSS_TIPS, PATH_DESIGN_SYSTEM, PATH_EXAMPLES, PATH_ROOT, PATH_SHOWCASE, PATH_TS_TIPS } from './path';
+import { CodeShowcase, Tips, DesignSystem, Examples, Home, UsefulLinks } from '../pages';
+import { PATH_TIPS, PATH_DESIGN_SYSTEM, PATH_EXAMPLES, PATH_ROOT, PATH_SHOWCASE, PATH_USEFUL_LINKS } from './path';
 import { Layout } from '../components/Layout';
 
 const router = createBrowserRouter([
@@ -11,8 +11,8 @@ const router = createBrowserRouter([
       { path: PATH_ROOT, element: <Home /> },
       { path: PATH_SHOWCASE, element: <CodeShowcase /> },
       { path: PATH_EXAMPLES, element: <Examples /> },
-      { path: PATH_CSS_TIPS, element: <CSSTips /> },
-      { path: PATH_TS_TIPS, element: <TSTips /> },
+      { path: PATH_TIPS, element: <Tips /> },
+      { path: PATH_USEFUL_LINKS, element: <UsefulLinks /> },
       { path: PATH_DESIGN_SYSTEM, element: <DesignSystem /> },
       { path: PATH_SHOWCASE, element: <CodeShowcase /> },
       { path: PATH_SHOWCASE, element: <CodeShowcase /> },
